--- server/components/validation/validationSchemas.js
@@ -1,42 +1,0 @@
-const { body } = require('express-validator/check');
-const mongoose = require('mongoose');
-
-exports.validate = (value) => {
-
-switch (value) {
-    case 'post-question': {
-     return [ 
-        body('type', 'unknown type').isString().matches(/question/),
-        body('title', 'Please provide a title')
-                      .not().isEmpty()
-                      .trim()
-                      .escape(),
-        body('body', 'Please provide a body')
-                      .not().isEmpty()
-                      .isLength({min:20}).withMessage('body must be at least 20 characters long')
-                      .trim()
-                      .escape(),
-        body('tags').optional().isArray()
-       ]
-    };
-    case 'post-reply': {
-      return [ 
-         body('type', 'unknown type').isString().matches(/reply/),
-         body('body', 'Please provide a body')
-                       .not().isEmpty()
-                       .isLength({min:20}).withMessage('body must be at least 20 characters long')
-                       .trim()
-                       .escape(),
-         body('authorId').not().isEmpty()
-                      .custom(value => {
-                        return mongoose.Types.ObjectId(value) == value
-                      })
-                      .customSanitizer(value => {
-                        return mongoose.Types.ObjectId(value)
-                      })
-        ]
-     };
-    default:
-      return;
-  }
-}
